# docs/openapi.yaml
openapi: 3.1.0
info:
  title: Simversity API
  version: 0.1.0
  description: AI-powered teaching simulator API for practicing responses to student misconceptions.
  contact:
    name: Simversity
    url: https://simversity.org
  license:
    name: MIT
servers:
  - url: https://simversity.org
    description: Production
  - url: http://localhost:3001
    description: Local development
...
paths:
...
  /api/conversations/{id}/observer:
    get:
      tags: [Observer]
      summary: Get observer messages
      security:
        - session: []
      parameters:
        - $ref: "#/components/parameters/id"
      responses:
        "200":
          description: Observer feedback messages
          content:
            application/json:
              schema:
                type: object
                properties:
                  messages:
                    type: array
                    items:
                      $ref: "#/components/schemas/ObserverMessage"
    post:
      tags: [Observer]
      summary: Generate observer feedback
      description: Streams observer feedback via SSE. Allowed on active and completed conversations.
      security:
        - session: []
      parameters:
        - $ref: "#/components/parameters/id"
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              required: [content]
              properties:
                content:
                  type: string
                  minLength: 1
                  maxLength: 5000
                  description: Teacher's question or prompt for the observer
      responses:
        "200":
          description: SSE stream of observer feedback
          content:
            text/event-stream:
              schema:
                $ref: "#/components/schemas/SSEEvent"
        "409":
          description: Cannot observe abandoned conversation
        "404":
          $ref: "#/components/responses/NotFound"
        "429":
          $ref: "#/components/responses/RateLimited"
...
components:
  securitySchemes:
    session:
      type: apiKey
      in: cookie
      name: better-auth.session_token
      description: Session cookie set by Better-Auth on sign-in
  parameters:
    id:
      name: id
      in: path
      required: true
      schema:
        type: string
        format: uuid
...
  schemas:
...
    ObserverMessage:
      type: object
      properties:
        id:
          type: string
          format: uuid
        role:
          type: string
          enum: [user, assistant]
        content:
          type: string
        sortOrder:
          type: integer
        createdAt:
          type: string
          format: date-time
...
    SSEEvent:
      description: |
        Server-Sent Event payload. All events use `event: message` with JSON data.
        The `type` field discriminates between event kinds.
      oneOf:
        - type: object
          title: Delta
          description: Streamed AI token chunk
          required: [type, text]
          properties:
            type:
              type: string
              enum: [delta]
            text:
              type: string
        - type: object
          title: Done
          description: Stream complete, response saved to DB
          required: [type, messageId]
          properties:
            type:
              type: string
              enum: [done]
            messageId:
              type: string
              format: uuid
            agentId:
              type: string
              format: uuid
              description: Present on conversation endpoints, absent on observer
            agentName:
              type: string
              description: Present on conversation endpoints, absent on observer
            truncated:
              type: boolean
              description: True when the model hit max_tokens
        - type: object
          title: Error
          description: AI failure or save failure
          required: [type, message]
          properties:
            type:
              type: string
              enum: [error]
            message:
              type: string
      discriminator:
        propertyName: type
...
    Error:
      type: object
      properties:
        error:
          type: string
  responses:
...
    NotFound:
      description: Resource not found
      content:
        application/json:
          schema:
            $ref: "#/components/schemas/Error"
...
    RateLimited:
      description: Rate limit exceeded
      content:
        application/json:
          schema:
            $ref: "#/components/schemas/Error"